Checkpoint no vessels clean
